Rename inmem_key_repo -> key_repo dir
