+++ test/core/libp2p/peer/key_book/inmem_key_repository_test.cpp
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 #include "libp2p/crypto/key.hpp"
-#include "libp2p/peer/inmem_key_repository/inmem_key_repository.hpp"
+#include "libp2p/peer/key_repository/inmem_key_repository.hpp"
 #include "libp2p/peer/key_repository.hpp"
 #include "testutil/outcome.hpp"
